requirements.txt
mysql-connector-python==8.0.26
beautifulsoup4==4.12.2
bs4==0.0.1
pdfminer.six==20221105
Pillow==9.5.0
